CareerInterests and CareerInterestsView converted to functional component
- Changed to functional components
- Restructured to follow MVC model

Tested and functional :smile:
