+++ services/frontend-v3/src/components/careerInterests/CareerInterestsView.jsx
@@ -3,8 +3,8 @@ import { FormattedMessage, injectIntl } from "react-intl";
 
 import { Row, Col, List, Tag, Typography } from "antd";
 
-class CareerInterestsView extends Component {
-  generateCareerInterestsInfoList(dataSource) {
+function CareerInterestsView(props) {
+  const generateCareerInterestsInfoList = dataSource => {
     return (
       <List
         itemLayout="horizontal"
@@ -16,9 +16,9 @@ class CareerInterestsView extends Component {
         )}
       />
     );
-  }
+  };
 
-  generateRelocationLocationsInfoList(dataSource) {
+  const generateRelocationLocationsInfoList = dataSource => {
     return (
       <div>
         <Typography.Text strong>
@@ -29,61 +29,18 @@ class CareerInterestsView extends Component {
         ))}
       </div>
     );
-  }
+  };
 
-  getCareerInterestsInfo(locale) {
-    const data = this.props.data;
-
-    const interestedInRemote = {
-      icon: "mail",
-      title: <FormattedMessage id="profile.interested.in.remote" />,
-      description:
-        data.interestedInRemote === true ? (
-          <FormattedMessage id="profile.yes" />
-        ) : (
-          <FormattedMessage id="profile.no" />
-        )
-    };
-    const lookingForNewJob = {
-      icon: "mail",
-      title: <FormattedMessage id="profile.looking.for.new.job" />,
-      description: (data.lookingForNewJob &&
-        data.lookingForNewJob.description[locale]) || (
-        <FormattedMessage id="profile.not.specified" />
-      )
-    };
-
-    return [interestedInRemote, lookingForNewJob];
-  }
-
-  getRelocationLocationsInfo(locale) {
-    const data = this.props.data;
-
-    const relocationLocationsInfo = [];
-    if (data.relocationLocations) {
-      data.relocationLocations.forEach(locationElement =>
-        relocationLocationsInfo.push(locationElement.description[locale])
-      );
-    }
-
-    return [...relocationLocationsInfo];
-  }
-
-  render() {
-    const locale = this.props.intl.formatMessage({ id: "language.code" });
-
-    const info = this.getCareerInterestsInfo(locale);
-    const relocationLocationsInfo = this.getRelocationLocationsInfo(locale);
-
-    return (
-      <Row>
-        <Col span={24}>
-          {this.generateCareerInterestsInfoList(info)}
-          {this.generateRelocationLocationsInfoList(relocationLocationsInfo)}
-        </Col>
-      </Row>
-    );
-  }
+  const info = props.info;
+  const relocationLocationsInfo = props.relocationLocationsInfo;
+  return (
+    <Row>
+      <Col span={24}>
+        {generateCareerInterestsInfoList(info)}
+        {generateRelocationLocationsInfoList(relocationLocationsInfo)}
+      </Col>
+    </Row>
+  );
 }
 
 export default injectIntl(CareerInterestsView);
